+++ modules/person.js
@@ -1,6 +1,6 @@
 const person = {
 	name: "Paweł Chabowski",
-	age: 32,
+	age: 32
 };
 
 export default person;
